"Added expense\_id to newExpense and updateExpense events in Expenses component"
This commit adds the `expense_id` field to the `newExpense` and `updateExpense` events in the `Expenses` component. In the backend controllers `Owners.js`, the `expense_id` is added to the data emitted for the `newExpense` event in the `createExpense` function and for the `updateExpense` event in the `updateExpense` function. The frontend component `Expenses.tsx` updates the `newExpense` and `updateExpense` event handlers to include the `expense_id` field in the `data` object.